requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.0
pmdarima>=2.0.0

# Deep Learning (Optional - install separately if needed)
# tensorflow>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# Dashboard
streamlit>=1.28.0
dash>=2.13.0
dash-bootstrap-components>=1.5.0

# Backend API
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Data Generation
faker>=19.0.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
requests>=2.31.0
APScheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.9.0
flake8>=6.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0
